backend/requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.6
django-environ==0.12.0
h11==0.14.0
psycopg==3.2.4
psycopg-binary==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
uvicorn==0.34.0
whitenoise==6.9.0